pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "redteam-local"
version = "0.1.0"
description = "Offline repo security auditor with Ollama-assisted remediation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0.1", "requests>=2.31.0"]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.5.0"]

[project.scripts]
rtl = "redteam_local.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
